+++ clickhouse_connect/driver/query.py
@@ -14,7 +14,7 @@ from clickhouse_connect.driver.common import dict_copy, empty_gen, StreamContext
 from clickhouse_connect.driver.external import ExternalData
 from clickhouse_connect.driver.types import Matrix, Closable
 from clickhouse_connect.driver.exceptions import StreamClosedError, ProgrammingError
-from clickhouse_connect.driver.options import check_arrow, pd_extended_dtypes
+from clickhouse_connect.driver.options import check_arrow
 from clickhouse_connect.driver.context import BaseQueryContext
 
 if TYPE_CHECKING:
@@ -128,7 +128,6 @@ class QueryContext(BaseQueryContext):
         self.response_tz = None
         self.block_info = False
         self.as_pandas = as_pandas
-        self.use_pandas_na = as_pandas and pd_extended_dtypes
         self.streaming = streaming
         self._rename_response_column: Optional[str] = rename_response_column
         self.column_renamer = get_rename_method(rename_response_column)